+++ core/src/main/java/com/landmuc/core/domain/model/UserInfo.kt
@@ -8,7 +8,7 @@ data class UserInfo(
     val gender: Gender,
     val activityLevel: ActivityLevel,
     val goalType: GoalType,
-    val carbRatio: Int,
-    val proteinRatio: Int,
-    val fatRatio: Int
+    val carbRatio: Float,
+    val proteinRatio: Float,
+    val fatRatio: Float
 )
